change score to a uint8 from 0 to 200

+++ pkg/common/query_result.go
@@ -1,7 +1,7 @@
 package common
 
 type QueryResult struct {
-	Score       float64
+	Score       uint8
 	Term        string
 	DictName    string
 	Definitions []string
